# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments/{assignmentId}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get an assignment with its problems
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{assignmentId}/start:
    post:
      operationId: startAssignmentAttempt
      tags: [assignments]
      summary: Start (or resume) an attempt on an assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptState"
  /assignments/attempts/{attemptId}:
    get:
      operationId: getAttempt
      tags: [assignments]
      summary: Get current attempt state (answers, time remaining)
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptState"
  /assignments/attempts/{attemptId}/answer:
    put:
      operationId: saveAnswer
      tags: [assignments]
      summary: Save (or update) a single answer with keystroke trace
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerSaved"
  /assignments/attempts/{attemptId}/submit:
    post:
      operationId: submitAttempt
      tags: [assignments]
      summary: Submit the attempt for grading (also runs AI + diachronic detection)
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResult"
  /assignments/{assignmentId}/readiness:
    get:
      operationId: getAssignmentReadiness
      tags: [assignments]
      summary: Analytics-based readiness pointers for an assignment (from logged practice)
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentReadiness"
...
components:
  schemas:
...
    Assignment:
      type: object
      required: [id, kind, title, weekNumber, isTimed, problems]
      properties:
        id: { type: integer }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        title: { type: string }
        weekNumber: { type: integer }
        isTimed: { type: boolean }
        timeLimitMinutes: { type: ["integer", "null"] }
        instructions: { type: ["string", "null"] }
        problems:
          type: array
          items:
            $ref: "#/components/schemas/Problem"
    Problem:
      type: object
      required: [id, position, prompt, topicId]
      properties:
        id: { type: integer }
        position: { type: integer }
        prompt: { type: string }
        topicId: { type: integer }
        topicTitle: { type: ["string", "null"] }
        hint: { type: ["string", "null"] }
    AttemptState:
      type: object
      required: [id, assignmentId, status, startedAt, answers]
      properties:
        id: { type: integer }
        assignmentId: { type: integer }
        status:
          type: string
          enum: [in_progress, submitted]
        startedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
        deadlineAt: { type: ["string", "null"], format: date-time }
        answers:
          type: array
          items:
            $ref: "#/components/schemas/SavedAnswer"
    SavedAnswer:
      type: object
      required: [problemId, answer]
      properties:
        problemId: { type: integer }
        answer: { type: string }
        keystrokeCount: { type: ["integer", "null"] }
        eraseCount: { type: ["integer", "null"] }
    AnswerInput:
      type: object
      required: [problemId, answer, trace]
      properties:
        problemId: { type: integer }
        answer: { type: string }
        trace:
          $ref: "#/components/schemas/KeystrokeTrace"
    KeystrokeTrace:
      type: object
      required: [keystrokeCount, eraseCount, durationMs]
      properties:
        keystrokeCount: { type: integer }
        eraseCount: { type: integer }
        bulkInsertCount: { type: integer, default: 0 }
        longestBulkInsertChars: { type: integer, default: 0 }
        rewriteSegments: { type: integer, default: 0 }
        durationMs: { type: integer }
    AnswerSaved:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    AttemptResult:
      type: object
      required: [attemptId, score, total, percent, perProblem, detection]
      properties:
        attemptId: { type: integer }
        score: { type: number }
        total: { type: number }
        percent: { type: number }
        perProblem:
          type: array
          items:
            $ref: "#/components/schemas/ProblemResult"
        detection:
          type: array
          items:
            $ref: "#/components/schemas/DetectionResult"
    ProblemResult:
      type: object
      required: [problemId, correct, explanation]
      properties:
        problemId: { type: integer }
        correct: { type: boolean }
        userAnswer: { type: string }
        correctAnswer: { type: string }
        explanation: { type: string }
    AssignmentReadiness:
      type: object
      required: [assignmentId, title, kind, readinessScore, verdict, practiceCount, topics, pointers]
      properties:
        assignmentId: { type: integer }
        title: { type: string }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        readinessScore:
          type: number
          description: 0-100 readiness estimate from practice accuracy + coverage
        verdict:
          type: string
          enum: [not_ready, getting_there, ready]
        practiceCount:
          type: integer
          description: Total practice attempts logged on this assignment's topics
        topics:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessTopic"
        pointers:
          type: array
          items: { type: string }
    ReadinessTopic:
      type: object
      required: [topicId, topicTitle, attempts, accuracy, strengthLabel]
      properties:
        topicId: { type: integer }
        topicTitle: { type: string }
        attempts: { type: integer }
        accuracy: { type: number }
        strengthLabel:
          type: string
          enum: [strong, solid, developing, weak, untested]
...
    DetectionResult:
      type: object
      required: [problemId, aiScore, aiFlagged, diachronicScore, diachronicFlagged, rationale]
      properties:
        problemId: { type: ["integer", "null"] }
        aiScore:
          type: number
          description: 0..1 probability the text is AI-generated
        aiFlagged: { type: boolean }
        diachronicScore:
          type: number
          description: 0..1 probability the user is rewording AI output (based on keystroke pattern)
        diachronicFlagged: { type: boolean }
        rationale: { type: string }
...