# lib/api-zod/src/generated/types/assignmentReadiness.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentReadinessKind } from './assignmentReadinessKind';
import type { AssignmentReadinessVerdict } from './assignmentReadinessVerdict';
import type { ReadinessTopic } from './readinessTopic';

export interface AssignmentReadiness {
  assignmentId: number;
  title: string;
  kind: AssignmentReadinessKind;
  /** 0-100 readiness estimate from practice accuracy + coverage */
  readinessScore: number;
  verdict: AssignmentReadinessVerdict;
  /** Total practice attempts logged on this assignment's topics */
  practiceCount: number;
  topics: ReadinessTopic[];
  pointers: string[];
}