# lib/api-zod/src/generated/types/readinessTopic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReadinessTopicStrengthLabel } from './readinessTopicStrengthLabel';

export interface ReadinessTopic {
  topicId: number;
  topicTitle: string;
  attempts: number;
  accuracy: number;
  strengthLabel: ReadinessTopicStrengthLabel;
}